pyproject.toml
[tool.poetry]
name = "status-bot"
version = "0.2.0"
description = ""
authors = ["Lonely-Dark <vlad10game@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.4"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
